+++ R/b_gff.R
@@ -21,11 +21,9 @@
 #'          c(22, 18, 24), c(23, 19, 25), c(24, 20)
 #'     )
 #'
-#'     m = b_gff(a, p = 4, symmetric = TRUE)
+#'     m = b_gff(a, p = 3, symmetric = TRUE)
 #'     image(matrix(m[, 1], 5, 5), col = pal)
-#'     image(matrix(m[, 2], 5, 5), col = pal)
 #'     image(matrix(m[, 3], 5, 5), col = pal)
-#'     image(matrix(m[, 4], 5, 5), col = pal)
 #'
 #'     if (requireNamespace("igraph", quietly = TRUE)) {
 #'         a = igraph::make_lattice(c(100, 100))
@@ -50,10 +48,10 @@ b_gff <- function(x, p = min(length(x) - 1L, 50), symmetric = FALSE) {
         sm = adj::adj_laplacian(x, sparse = TRUE)
     }
     if (p >= length(x)) {
-        rlang::abort("Cannot generate more features than the number of nodes minus one.")
+        abort("Cannot generate more features than the number of nodes minus one.")
     }
     if (p <= 0) {
-        rlang::abort("`p` must be a positive integer.")
+        abort("`p` must be a positive integer.")
     }
 
     fn = if (isTRUE(symmetric)) RSpectra::eigs_sym else RSpectra::eigs