+++ NEWS.md
@@ -6,6 +6,7 @@
 * New `b_rocket()` for random convolutional features for regression on time series TODO https://arxiv.org/pdf/2012.08791
 * New `b_echo()` for echo state network features for time series forecasting TODO https://www.ai.rug.nl/minds/uploads/PracticalESN.pdf
 * More efficient `b_ker()` option for many predictions
+* Control automatic leaf pruning in `b_bart()`
 * New vignette on other packages that help produce basis expansions or embeddings.
 
 # bases 0.1.2