+++ NEWS.md
@@ -6,7 +6,7 @@
 * New `b_gff()` for graph Fourier features for regression on spatial and graph-structured data.
   Requires `RSpectra` package for efficient eigendecomposition, and either
   `adj` or `igraph` for graph representation.
-* New `b_conv()` for random convolutional features for regression on images TODO im2col()
+* New `b_conv()` for random convolutional features for regression on images
 * New `b_rocket()` for random convolutional features for regression on time series TODO https://arxiv.org/pdf/2012.08791
 * New `b_echo()` for echo state network features for time series forecasting TODO https://www.ai.rug.nl/minds/uploads/PracticalESN.pdf
 * More efficient `b_ker()` option for many predictions